modification of the noise array calculation to reduce CPU time
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@5820 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ ecl/modules/eclDigitizer/include/ECLDigitizerModule.h
@@ -145,7 +145,7 @@ namespace Belle2 {
       /**input array to fit          */
       int FitA[31];
       /** array for calculation of random noise for FitA[31]      */
-      double m_vmat[31][31];
+      float m_vmat[31][31];
 
     };
   }//ECL
